comment deletion fix. mod powers on comments

+++ src/components/Post/Comment.js
@@ -17,6 +17,8 @@ function Comment(props) {
     });
     const [commentId, setCommentId] = useState(null);
 
+    let deleteStarted = 0;
+
     const popoverStyle = {
         backgroundColor: "#292833",
         boxShadow: "4px 4px 25px 0px #171421 ",
@@ -37,33 +39,61 @@ function Comment(props) {
         props.getUser(comment.author, setAuthor);
     }, [comment])
 
-    const popover = (
-        <Popover id={commentId + "OptionsPopover"} style={popoverStyle}>
-            <Popover.Content>
-                <div className="row">
-                    <p onClick={() => handleDeleteClicked(commentId, comment.postId)} style={{color: "#ffffff"}}>Delete</p>
-                </div>
-            </Popover.Content>
-        </Popover>
-    );
-
-    const checkOptions = () => {
-        if (props.currentUserId === comment.author) {
-            return (<OverlayTrigger trigger="click" rootClose placement="bottom" overlay={popover}>
-                <img src={optionsIcon} style={{
-                    backgroundColor: "transparent",
-                    position: "relative",
-                    bottom: "-8px",
-                    left: "100px"
-                }} />
-            </OverlayTrigger>)
+    //create maping of emote with a list of the users who reacted
+  function reactionMap() {
+    let map = new Map();
+    const reacted = comment.reacted;
+    if (reacted) {
+      for (const user in reacted) {
+        const reaction = reacted[user];
+        if (map.has(reaction)) {
+          map.get(reaction).push(user);
+        } else {
+          map.set(reaction, [user])
         }
+      }
     }
+    return map
+  }
+
+    function handleShowReactions() {
+        //show the modal and then set the current reactions
+        props.setShowModal(true);
+        //calculate the reaction mapping and set it to modal content
+        props.setModalContent(reactionMap());
+      }
+
+    function checkOptions() {
+        let modLvl;
+        if (!props.currentUserInfo.moderationLevel) {
+            modLvl = 0;
+        } else {
+            modLvl = props.currentUserInfo.moderationLevel;
+        }
+        return (
+            <div>
+                <div id="dropdownMenuButton" className="btn" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" style={{ background: "transparent" }}>
+                    <img src={optionsIcon} alt={"options"} style={{ backgroundColor: "transparent" }} />
+                </div>
+                <div className="dropdown-menu-right dropdown-menu" aria-labelledby="dropdownMenuButton">
+                    <p className="dropdown-item mb-0" onClick={() => handleDeleteClicked(commentId, comment.postId)} style={{ cursor: "pointer" }}>Delete Comment</p>
+                    <p className="dropdown-item mb-0" onClick={() => handleShowReactions()} style={{ cursor: "pointer" }}>Reactions</p>
+                    <p className="dropdown-item mb-0" onClick={() => props.report("content/comments", commentId)} style={{ cursor: "pointer" }}>Report Comment</p>
+                    {modLvl > 0 ? <p className="dropdown-item mb-0" onClick={() => props.clearReports("content/comments", commentId)} style={{ cursor: "pointer" }}>Clear Reports (Current: {comment.reports ? comment.reports : 0})</p> : null}
+                </div>
+            </div>
+        )
+    };
 
     const handleDeleteClicked = (tempCommentId, tempPostId) => {
-        props.deleteComment(tempCommentId, tempPostId);
-        document.body.click();
-        props.updateRefresh();
+        if (deleteStarted === 0) {
+            deleteStarted++;
+            if (deleteStarted === 1) {
+                props.deleteComment(tempCommentId, tempPostId);
+                document.body.click();
+                props.updateRefresh();
+            }
+        }
     }
 
     return (
@@ -94,7 +124,7 @@ function Comment(props) {
                 </div>
                 <div className="row">
                     <div className="col-12">
-                        <Text text={comment.commentText} style={{whiteSpace: "pre-wrap"}}/>
+                        <Text text={comment.commentText} style={{ whiteSpace: "pre-wrap" }} />
                     </div>
                 </div>
                 <div className="row">
